refactor(web-main): replace install versionId with tag
