+++ packages/web-main/src/routes/_auth/gameserver.$gameServerId/modules.$moduleId.$moduleVersionTag.install.view.tsx
@@ -3,14 +3,17 @@ import { moduleQueryOptions } from '../../../queries/module';
 import { createFileRoute, redirect } from '@tanstack/react-router';
 import { InstallModuleForm } from './-InstallModuleForm';
 import { hasPermission } from '../../../hooks/useHasPermission';
+import { PERMISSIONS } from '@takaro/apiclient';
 import { userMeQueryOptions } from '../../../queries/user';
 import { DrawerSkeleton } from '@takaro/lib-components';
 import { useQueries } from '@tanstack/react-query';
 
-export const Route = createFileRoute('/_auth/gameserver/$gameServerId/modules/$moduleId/$versionId/update')({
+export const Route = createFileRoute(
+  '/_auth/gameserver/$gameServerId/modules/$moduleId/$moduleVersionTag/install/view',
+)({
   beforeLoad: async ({ context }) => {
     const session = await context.queryClient.ensureQueryData(userMeQueryOptions());
-    if (!hasPermission(session, ['READ_MODULES', 'MANAGE_MODULES'])) {
+    if (!hasPermission(session, [PERMISSIONS.ReadModules])) {
       throw redirect({ to: '/forbidden' });
     }
   },
@@ -38,6 +41,7 @@ function Component() {
 
   return (
     <InstallModuleForm
+      readOnly={true}
       gameServerId={gameServerId}
       modVersion={modInstallation.version}
       modInstallation={modInstallation}
